doc(.github/ISSUE_TEMPLATE): simplify bug report template (#48)
Simplify the bug report template by removing the Additional
Information/Context, OS, and Server Version sections for now.

Move the server dropdown to the top of the template, since it’s the
first piece of information needed for triaging.

Leave the title field empty so users can’t accidentally submit it with
the default `package: short issue description`.