chore: delete .eslintrc.json (#63)
This will ensure that we only use eslintrc.config.js, which will ensure
header checking.

Unfortunately this library is using a newer eslint config which gts is
not up to date with. So, we will skip importing the gts eslint config
rules for now, and I'll make a TODO issue to update this config once
gts' eslint is up-to-date.

see https://github.com/googleapis/gcloud-mcp/issues/64