fix: server disconnected (#72)
fix: server disconnected

Previously when running `gemini mcp list` the gcloud server would say
disconnect. gemini actually tries to start the server with this command
and the reason it wasn't connecting is because the server expects a
gcloud executable to be available at startup. Now the integration test
install gcloud CLI and shows connected during `gemini mcp list`. Also
updated the assertion to ensure its connected and not just listed.