fix: disallow gcloud interactive, as it will fail (#88)
Local testing of having the tool run with `gcloud beta interactive`
showed that the tool doesn't return.

Unsurprising result, given there's no way for the user to interact with
the tool, gcloud is expecting interaction, and gcloud can't just fall
back to a default.

Added `interactive` to the denylist to prevent this freezing, and added
a note in the tool description to proactively avoid the
denylisted-and-will-fail commands.