fix: update gemini cli extension name (#164)
As an initial fix for
https://github.com/google-gemini/gemini-cli/pull/7065 to allow the
migration to succeed, the extension name of the mcp servers are
standardized to not have directory (slash) notation in the extension
name. This causes issue when gemini CLI goes to find the extensions.json
file.

To test this on a clean worksapce:
1) Install nightly gemini-cli -- `npm install -g
@google/gemini-cli@nightly`
2) init an mcp server a project level dir: `gcloud-mcp init
--agent=gemini-cli --local`
3) Run `gemini` then accept prompt to migrate
4) rerun `gemini` and mcp server should be working with no extension
warning at gemini start.


A follow up PR to this may include removing the workspace-level install
in favor of the root / home level `/gemini/extensions` install.