fix: standardize install to home dir (#178)
With the introduction of
https://github.com/google-gemini/gemini-cli/pull/7065 -- Gemini CLI's
new guidance is to install extensions at the user/home level instead of
at the project/workspace level. This still works well with our local
development flow as the bin command is available at that level.

If we do not include this change, users of Gemini CLI will be prompted
to migrate their extensions to home dir.